Start developing against AssertJ 4.x and make sure the library is a JPMS module.
