+++ src/main/java/module-info.java
@@ -1,0 +1,25 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * Copyright 2025-2025 the original author or authors.
+ */
+
+/**
+ * Main AssertJ Eclipse Collections Module
+ */
+module org.assertj.eclipse.collections {
+  exports org.assertj.eclipse.collections.api;
+  exports org.assertj.eclipse.collections.api.multimap;
+  exports org.assertj.eclipse.collections.error;
+
+  requires org.assertj.core;
+  requires org.eclipse.collections.api;
+  requires org.eclipse.collections.impl;
+}
