+++ src/test/java/org/assertj/eclipse/collections/test/api/multimap/MultimapAssert_ContainsOnly_Test.java
@@ -10,7 +10,7 @@
  *
  * Copyright 2025-2025 the original author or authors.
  */
-package org.assertj.eclipse.collections.api.multimap;
+package org.assertj.eclipse.collections.test.api.multimap;
 
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.assertj.core.api.Assertions.assertThatNoException;
@@ -19,6 +19,7 @@ import static org.eclipse.collections.impl.tuple.Tuples.pair;
 import java.util.Map;
 
 import org.assertj.eclipse.collections.api.SoftAssertions;
+import org.assertj.eclipse.collections.api.multimap.MultimapAssert;
 import org.eclipse.collections.api.multimap.Multimap;
 import org.eclipse.collections.api.tuple.Pair;
 import org.junit.jupiter.api.Test;
@@ -28,7 +29,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 class MultimapAssert_ContainsOnly_Test {
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#shipMultimaps")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#shipMultimaps")
   void passesWithPairs(Multimap<String, String> actual) {
     Pair<String, String>[] exactMatchPairs = new Pair[]{
       pair("TNG", "Enterprise"),
@@ -40,7 +41,7 @@ class MultimapAssert_ContainsOnly_Test {
   }
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#shipMultimaps")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#shipMultimaps")
   void passesWithEntries(Multimap<String, String> actual) {
     Map.Entry<String, String>[] exactMatchEntries = new Map.Entry[]{
       pair("TNG", "Enterprise").toEntry(),
@@ -52,7 +53,7 @@ class MultimapAssert_ContainsOnly_Test {
   }
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#shipMultimaps")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#shipMultimaps")
   void failsWhenAdditionalKeysOrValuesExistWithPair(Multimap<String, String> actual) {
     Pair<String, String>[] partialMatchMissingPairs = new Pair[]{
       pair("TNG", "Enterprise"),
@@ -66,7 +67,7 @@ class MultimapAssert_ContainsOnly_Test {
   }
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#shipMultimaps")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#shipMultimaps")
   void failsWhenAdditionalKeysOrValuesExistWithEntry(Multimap<String, String> actual) {
     Map.Entry<String, String>[] partialMatchMissingEntries = new Map.Entry[]{
       pair("TNG", "Enterprise").toEntry(),
@@ -80,7 +81,7 @@ class MultimapAssert_ContainsOnly_Test {
   }
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#shipMultimaps")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#shipMultimaps")
   void failsWhenEntryIsMissingWithPair(Multimap<String, String> actual) {
     Pair<String, String>[] partialMatchExtraPairs = new Pair[]{
       pair("TOS", "Enterprise"),
@@ -96,7 +97,7 @@ class MultimapAssert_ContainsOnly_Test {
   }
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#shipMultimaps")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#shipMultimaps")
   void failsWhenEntryIsMissingWithEntry(Multimap<String, String> actual) {
     Map.Entry<String, String>[] partialMatchExtraEntries = new Map.Entry[]{
       pair("TOS", "Enterprise").toEntry(),
@@ -112,7 +113,7 @@ class MultimapAssert_ContainsOnly_Test {
   }
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#emptyMultimaps")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#emptyMultimaps")
   void failsForEmptyMultimapWithPair(Multimap<String, String> actual) {
     Pair<String, String>[] exactMatchPairs = new Pair[]{
       pair("TNG", "Enterprise"),
@@ -125,7 +126,7 @@ class MultimapAssert_ContainsOnly_Test {
   }
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#emptyMultimaps")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#emptyMultimaps")
   void failsForEmptyMultimapWithEntry(Multimap<String, String> actual) {
     Map.Entry<String, String>[] exactMatchEntries = new Map.Entry[]{
       pair("TNG", "Enterprise").toEntry(),
@@ -160,7 +161,7 @@ class MultimapAssert_ContainsOnly_Test {
   }
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#shipMultimaps")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#shipMultimaps")
   void softAssertionPassesWithPairs(Multimap<String, String> actual) {
     Pair<String, String>[] exactMatchPairs = new Pair[]{
       pair("TNG", "Enterprise"),
@@ -172,7 +173,7 @@ class MultimapAssert_ContainsOnly_Test {
   }
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#shipMultimaps")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#shipMultimaps")
   void softAssertionPassesWithEntries(Multimap<String, String> actual) {
     Map.Entry<String, String>[] exactMatchEntries = new Map.Entry[]{
       pair("TNG", "Enterprise").toEntry(),
