+++ src/test/java/org/assertj/eclipse/collections/test/api/multimap/MultimapAssert_HasDistinctSizeGreaterThan_Test.java
@@ -10,14 +10,14 @@
  *
  * Copyright 2025-2025 the original author or authors.
  */
-package org.assertj.eclipse.collections.api.multimap;
+package org.assertj.eclipse.collections.test.api.multimap;
 
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
 import org.assertj.eclipse.collections.api.SoftAssertions;
+import org.assertj.eclipse.collections.api.multimap.MultimapAssert;
 import org.eclipse.collections.api.multimap.Multimap;
-import org.eclipse.collections.impl.factory.Multimaps;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -25,13 +25,13 @@ import org.junit.jupiter.params.provider.MethodSource;
 class MultimapAssert_HasDistinctSizeGreaterThan_Test {
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#distinctSizeLowerBoundaryTestData")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#distinctSizeLowerBoundaryTestData")
   void passesGreaterThan(Multimap<String, String> actual, int lowerBoundary) {
     assertThatNoException().isThrownBy(() -> new MultimapAssert<>(actual).hasDistinctSizeGreaterThan(lowerBoundary));
   }
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#distinctSizeUpperBoundaryTestData")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#distinctSizeUpperBoundaryTestData")
   void failsLesser(Multimap<String, String> actual, int upperBoundary) {
     assertThatExceptionOfType(AssertionError.class)
       .isThrownBy(() -> new MultimapAssert<>(actual).hasDistinctSizeGreaterThan(upperBoundary))
@@ -40,7 +40,7 @@ class MultimapAssert_HasDistinctSizeGreaterThan_Test {
   }
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#distinctSizeEqualsTestData")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#distinctSizeEqualsTestData")
   void failsEquals(Multimap<String, String> actual, int equalsBoundary) {
     assertThatExceptionOfType(AssertionError.class)
       .isThrownBy(() -> new MultimapAssert<>(actual).hasDistinctSizeGreaterThan(equalsBoundary))
@@ -57,7 +57,7 @@ class MultimapAssert_HasDistinctSizeGreaterThan_Test {
   }
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#emptyMultimaps")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#emptyMultimaps")
   void failsEmptyMultimap(Multimap<String, String> actual) {
     int lowerBoundary = 2; // Using the same value as in distinctSizeLowerBoundaryTestData
     assertThatExceptionOfType(AssertionError.class)
@@ -66,7 +66,7 @@ class MultimapAssert_HasDistinctSizeGreaterThan_Test {
   }
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#distinctSizeLowerBoundaryTestData")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#distinctSizeLowerBoundaryTestData")
   void softAssertionPasses(Multimap<String, String> actual, int lowerBoundary) {
     SoftAssertions.assertSoftly(softly -> softly.assertThat(actual).hasDistinctSizeGreaterThan(lowerBoundary));
   }
