+++ src/test/java/org/assertj/eclipse/collections/test/api/multimap/MultimapAssert_HasSizeLessThan_Test.java
@@ -10,50 +10,54 @@
  *
  * Copyright 2025-2025 the original author or authors.
  */
-package org.assertj.eclipse.collections.api.multimap;
+package org.assertj.eclipse.collections.test.api.multimap;
 
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
 import org.assertj.eclipse.collections.api.SoftAssertions;
+import org.assertj.eclipse.collections.api.multimap.MultimapAssert;
 import org.eclipse.collections.api.multimap.Multimap;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class MultimapAssert_HasSizeLessThanOrEqualTo_Test {
+class MultimapAssert_HasSizeLessThan_Test {
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#sizeUpperBoundaryTestData")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#sizeUpperBoundaryTestData")
   void passesLessThan(Multimap<String, String> actual, int upperBoundary) {
-    assertThatNoException().isThrownBy(() -> new MultimapAssert<>(actual).hasSizeLessThanOrEqualTo(upperBoundary));
+    assertThatNoException().isThrownBy(() -> new MultimapAssert<>(actual).hasSizeLessThan(upperBoundary));
   }
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#sizeEqualsTestData")
-  void passesEqual(Multimap<String, String> actual, int equalsBoundary) {
-    assertThatNoException().isThrownBy(() -> new MultimapAssert<>(actual).hasSizeLessThanOrEqualTo(equalsBoundary));
-  }
-
-  @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#sizeLowerBoundaryTestData")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#sizeLowerBoundaryTestData")
   void failsGreater(Multimap<String, String> actual, int lowerBoundary) {
     assertThatExceptionOfType(AssertionError.class)
-      .isThrownBy(() -> new MultimapAssert<>(actual).hasSizeLessThanOrEqualTo(lowerBoundary))
+      .isThrownBy(() -> new MultimapAssert<>(actual).hasSizeLessThan(lowerBoundary))
       .withMessageContaining("Expecting size of")
-      .withMessageContaining(String.format("to be less than or equal to %s but was %s", lowerBoundary, actual.size()));
+      .withMessageContaining(String.format("to be less than %s but was %s", lowerBoundary, actual.size()));
+  }
+
+  @ParameterizedTest
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#sizeEqualsTestData")
+  void failsEquals(Multimap<String, String> actual, int equalsBoundary) {
+    assertThatExceptionOfType(AssertionError.class)
+      .isThrownBy(() -> new MultimapAssert<>(actual).hasSizeLessThan(equalsBoundary))
+      .withMessageContaining("Expecting size of")
+      .withMessageContaining(String.format("to be less than %s but was %s", equalsBoundary, actual.size()));
   }
 
   @Test
   void failsNullMultimap() {
     assertThatExceptionOfType(AssertionError.class)
-      .isThrownBy(() -> new MultimapAssert<>(null).hasSizeLessThanOrEqualTo(42))
+      .isThrownBy(() -> new MultimapAssert<>(null).hasSizeLessThan(50))
       .withMessageContaining("Expecting actual not to be null");
   }
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#sizeUpperBoundaryTestData")
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#sizeUpperBoundaryTestData")
   void softAssertionPasses(Multimap<String, String> actual, int upperBoundary) {
-    SoftAssertions.assertSoftly(softly -> softly.assertThat(actual).hasSizeLessThanOrEqualTo(upperBoundary));
+    SoftAssertions.assertSoftly(softly -> softly.assertThat(actual).hasSizeLessThan(upperBoundary));
   }
 }
