+++ src/test/java/org/assertj/eclipse/collections/test/api/multimap/MultimapAssert_IsNotEmpty_Test.java
@@ -10,43 +10,44 @@
  *
  * Copyright 2025-2025 the original author or authors.
  */
-package org.assertj.eclipse.collections.api.multimap;
+package org.assertj.eclipse.collections.test.api.multimap;
 
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
 import org.assertj.eclipse.collections.api.SoftAssertions;
+import org.assertj.eclipse.collections.api.multimap.MultimapAssert;
 import org.eclipse.collections.api.multimap.Multimap;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class MultimapAssert_HasSize_Test {
+class MultimapAssert_IsNotEmpty_Test {
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#sizeEqualsTestData")
-  void passes(Multimap<String, String> actual, int expectedSize) {
-    assertThatNoException().isThrownBy(() -> new MultimapAssert<>(actual).hasSize(expectedSize));
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#nonEmptyMultimaps")
+  void passesNotEmpty(Multimap<String, String> actual) {
+    assertThatNoException().isThrownBy(() -> new MultimapAssert<>(actual).isNotEmpty());
   }
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#emptyMultimapsWithExpectedSize")
-  void failsEmpty(Multimap<String, String> actual, int expectedSize) {
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#emptyMultimaps")
+  void failsEmpty(Multimap<String, String> actual) {
     assertThatExceptionOfType(AssertionError.class)
-      .isThrownBy(() -> new MultimapAssert<>(actual).hasSize(expectedSize))
-      .withMessageContaining(String.format("Expected size: %s but was: 0", expectedSize));
+      .isThrownBy(() -> new MultimapAssert<>(actual).isNotEmpty())
+      .withMessageContaining("Expecting actual not to be empty");
   }
 
   @Test
   void failsNullMultimap() {
     assertThatExceptionOfType(AssertionError.class)
-      .isThrownBy(() -> new MultimapAssert<>(null).hasSize(38))
+      .isThrownBy(() -> new MultimapAssert<>(null).isNotEmpty())
       .withMessageContaining("Expecting actual not to be null");
   }
 
   @ParameterizedTest
-  @MethodSource("org.assertj.eclipse.collections.api.multimap.MultimapTestData#sizeEqualsTestData")
-  void softAssertionPasses(Multimap<String, String> actual, int expectedSize) {
-    SoftAssertions.assertSoftly(softly -> softly.assertThat(actual).hasSize(expectedSize));
+  @MethodSource("org.assertj.eclipse.collections.test.api.multimap.MultimapTestData#nonEmptyMultimaps")
+  void softAssertionPasses(Multimap<String, String> actual) {
+    SoftAssertions.assertSoftly(softly -> softly.assertThat(actual).isNotEmpty());
   }
 }
